Extract capacity SQL builder into src/lib/db/capacity.ts
Move buildCapacityCondition and its dateToRange helper out of attendees.ts
into a new, standalone module so the capacity-check logic has a clear home
for future extension (e.g. multi-day bookings). attendees.ts now imports
both back in; events.ts imports dateToRange from capacity.ts directly.
